chore: split up cart page
